add mike's suggestions
