+++ src/PresentationalComponents/LoadingState/index.tsx
@@ -1,0 +1,3 @@
+import LoadingState from './LoadingState';
+
+export default LoadingState;
